# api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipelineRunIn": {
        "properties": {
          "ciclo": {
            "title": "Ciclo",
            "type": "string"
          },
          "dag": {
            "title": "Dag",
            "type": "string"
          }
        },
        "required": [
          "dag",
          "ciclo"
        ],
        "title": "PipelineRunIn",
        "type": "object"
      },
      "PipelineRunOut": {
        "properties": {
          "estado": {
            "default": "accepted",
            "title": "Estado",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id"
        ],
        "title": "PipelineRunOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Contrato REST de FARO (Escuela como Sensor Social). Fuente de verdad: 03_Architecture/API_Specification.md. Este servicio publica el OpenAPI que consumen los mocks de las Células 2 y 3 (US-401).",
    "title": "FARO API — Contrato v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/export": {
      "get": {
        "description": "Exporta datos en bruto (rol: **analista**). En el stub devuelve una referencia, no el stream.",
        "operationId": "export_api_v1_admin_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "tabla",
            "required": true,
            "schema": {
              "title": "Tabla",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ciclo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ciclo"
            }
          },
          {
            "in": "query",
            "name": "formato",
            "required": false,
            "schema": {
              "default": "csv",
              "title": "Formato",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Api V1 Admin Export Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Export",
        "tags": [
          "Administración"
        ]
      }
    },
...
    "/api/v1/admin/pipeline/run": {
      "post": {
        "description": "Encola una corrida real de `body.dag` en Airflow (rol: **analista**).\n\n`dag` se valida contra `DAGS_VALIDOS` (los DAG reales de `dags/`) -- 422 si no está en la\nlista, nunca se le pasa texto libre a Airflow.",
        "operationId": "pipeline_run_api_v1_admin_pipeline_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Pipeline Run",
        "tags": [
          "Administración"
        ]
      }
    },
...
  }
}